+++ RoR2_ItemsMod/Modules/Equipment/RespawnFlag.cs
@@ -304,6 +304,7 @@ namespace ExtradimensionalItems.Modules.Equipment
         {
             LoadAssetBundle();
             LoadSoundBank();
+            SetLogbookCameraPosition();
             LoadInteractable();
             CreateConfig(config);
             LoadLanguageFile();