+++ RoR2_ItemsMod/Modules/Equipment/SkullOfDoom.cs
@@ -341,6 +341,7 @@ namespace ExtradimensionalItems.Modules.Equipment
         {
             LoadAssetBundle();
             LoadSoundBank();
+            SetLogbookCameraPosition();
             CreateConfig(config);
             LoadLanguageFile();
             CreateBuffs();
@@ -469,11 +470,6 @@ namespace ExtradimensionalItems.Modules.Equipment
             ContentAddition.AddBuffDef(SkullOfDoomBuff);
 
             Content.Buffs.SkullOfDoom = SkullOfDoomBuff;
-
-            if (BetterUICompat.enabled)
-            {
-                BetterUICompat.AddBuffInfo(SkullOfDoomBuff, "BUFF_SKULL_OF_DOOM_NAME", "BUFF_SKULL_OF_DOOM_DESCRIPTION");
-            }
         }
 
         private void CreateVisualEffects()