+++ RoR2_ItemsMod/Modules/Interactables/RespawnFlagInteractableManager.cs
@@ -172,6 +172,10 @@ namespace ExtradimensionalItems.Modules.Interactables
             return false;
         }
 
+        public bool ShouldProximityHighlight()
+        {
+            return true;
+        }
     }
 
 }