+++ RoR2_ItemsMod/Modules/Items/DamageOnCooldowns.cs
@@ -345,6 +345,7 @@ namespace ExtradimensionalItems.Modules.Items
         public override void Init(ConfigFile config)
         {
             LoadAssetBundle();
+            SetLogbookCameraPosition();
             CreateConfig(config);
             LoadLanguageFile();
             CreateBuffs();
@@ -364,11 +365,6 @@ namespace ExtradimensionalItems.Modules.Items
             ContentAddition.AddBuffDef(DamageOnCooldownsBuff);
 
             Content.Buffs.DamageOnCooldowns = DamageOnCooldownsBuff;
-
-            if (BetterUICompat.enabled)
-            {
-                BetterUICompat.AddBuffInfo(DamageOnCooldownsBuff, "BUFF_DAMAGE_ON_COOLDOWNS_NAME", "BUFF_DAMAGE_ON_COOLDOWNS_DESCRIPTION");
-            }
         }
 
         protected override void Hooks()
@@ -422,16 +418,6 @@ namespace ExtradimensionalItems.Modules.Items
                 body.RemoveBuff(Content.Buffs.DamageOnCooldowns);
                 count++;
             }
-        }
-
-        public override void AddBetterUIStats(ItemDef item)
-        {
-            BetterUICompat.RegisterStat(item, "BETTERUICOMPAT_DESC_DAMAGE", DamageBonus.Value / 100, DamageBonusPerStack.Value / 100, BetterUICompat.StackingFormula.Linear, BetterUICompat.StatFormatter.Percent);
-        }
-
-        protected override void ModifyBetterUIStats()
-        {
-            BetterUICompat.ModifyBetterUIStat(ItemDef, "BETTERUICOMPAT_DESC_DAMAGE", DamageBonus.Value / 100, DamageBonusPerStack.Value / 100);
         }
 
         public override void CreateConfig(ConfigFile config)