+++ RoR2_ItemsMod/Modules/Items/Sheen.cs
@@ -289,6 +289,7 @@ namespace ExtradimensionalItems.Modules.Items
         {
             CreateConfig(config);
             LoadAssetBundle();
+            SetLogbookCameraPosition();
             CreateBuffs();
             CreateItem(ref Content.Items.Sheen);
             LoadLanguageFile();
@@ -337,12 +338,6 @@ namespace ExtradimensionalItems.Modules.Items
             ContentAddition.AddBuffDef(SheenBuff);
 
             Content.Buffs.Sheen = SheenBuff;
-
-            if (BetterUICompat.enabled)
-            {
-                BetterUICompat.AddBuffInfo(SheenBuff, "BUFF_SHEEN_NAME", "BUFF_SHEEN_DESCRIPTION");
-            }
-
         }
 
         private void CreateVisualEffects()
@@ -360,18 +355,6 @@ namespace ExtradimensionalItems.Modules.Items
 
             TempVisualEffectAPI.AddTemporaryVisualEffect(SheenEffectInstance.InstantiateClone("SheenEffectL", false), (CharacterBody body) => { return body.HasBuff(Content.Buffs.Sheen); }, true, "HandL");
             TempVisualEffectAPI.AddTemporaryVisualEffect(SheenEffectInstance.InstantiateClone("SheenEffectR", false), (CharacterBody body) => { return body.HasBuff(Content.Buffs.Sheen); }, true, "HandR");
-        }
-
-        public override void AddBetterUIStats(ItemDef item)
-        {
-            BetterUICompat.RegisterStat(item, "BETTERUICOMPAT_DESC_DAMAGE", DamageModifier.Value / 100, DamageModifierPerStack.Value / 100, BetterUICompat.StackingFormula.Linear, BetterUICompat.StatFormatter.Percent);
-            BetterUICompat.RegisterStat(item, "BETTERUICOMPAT_DESC_STACKS", BuffStackPerItem.Value, BuffStackPerItemPerStack.Value, BetterUICompat.StackingFormula.Linear, BetterUICompat.StatFormatter.Charges);
-        }
-
-        protected override void ModifyBetterUIStats()
-        {
-            BetterUICompat.ModifyBetterUIStat(ItemDef, "BETTERUICOMPAT_DESC_DAMAGE", DamageModifier.Value / 100, DamageModifierPerStack.Value / 100);
-            BetterUICompat.ModifyBetterUIStat(ItemDef, "BETTERUICOMPAT_DESC_STACKS", BuffStackPerItem.Value, BuffStackPerItemPerStack.Value);
         }
 
         public override void CreateConfig(ConfigFile config)