+++ RoR2_ItemsMod/Modules/Items/VoidCooldownReduction.cs
@@ -293,6 +293,7 @@ namespace ExtradimensionalItems.Modules.Items
         {
             CreateConfig(config);
             LoadAssetBundle();
+            SetLogbookCameraPosition();
             LoadLanguageFile();
             CreateItem(ref Content.Items.VoidCooldownReduction);
             Hooks();
@@ -314,16 +315,6 @@ namespace ExtradimensionalItems.Modules.Items
             };
             ItemCatalog.itemRelationships[DLC1Content.ItemRelationshipTypes.ContagiousItem] = ItemCatalog.itemRelationships[DLC1Content.ItemRelationshipTypes.ContagiousItem].AddToArray(transformation);
             orig();
-        }
-
-        public override void AddBetterUIStats(ItemDef item)
-        {
-            BetterUICompat.RegisterStat(item, "BETTERUICOPMAT_COOLDOWN_REDUCTUION", CooldownReduction.Value, BetterUICompat.StackingFormula.ProbablyExponential, BetterUICompat.StatFormatter.Percent, BetterUICompat.ItemTag.CooldownReduction);
-        }
-
-        protected override void ModifyBetterUIStats()
-        {
-            BetterUICompat.ModifyBetterUIStat(ItemDef, "BETTERUICOPMAT_COOLDOWN_REDUCTUION", CooldownReduction.Value, CooldownReduction.Value);
         }
 
         private void RecalculateStatsAPI_GetStatCoefficients(CharacterBody sender, RecalculateStatsAPI.StatHookEventArgs args)